test: migrate cwd tests to vitest
